requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
beautifulsoup4==4.13.4
    # via -r requirements.in
certifi==2025.7.14
    # via requests
charset-normalizer==3.4.2
    # via requests
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
et-xmlfile==2.0.0
    # via openpyxl
fonttools==4.59.0
    # via matplotlib
idna==3.10
    # via requests
joblib==1.5.1
    # via scikit-learn
kiwisolver==1.4.8
    # via matplotlib
lxml==6.0.0
    # via -r requirements.in
matplotlib==3.10.3
    # via -r requirements.in
numpy==2.3.1
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
openpyxl==3.1.5
    # via -r requirements.in
packaging==25.0
    # via matplotlib
pandas==2.3.1
    # via -r requirements.in
pillow==11.3.0
    # via matplotlib
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via -r requirements.in
requests==2.32.4
    # via -r requirements.in
scikit-learn==1.7.1
    # via -r requirements.in
scipy==1.16.0
    # via scikit-learn
six==1.17.0
    # via python-dateutil
soupsieve==2.7
    # via beautifulsoup4
threadpoolctl==3.6.0
    # via scikit-learn
typing-extensions==4.14.1
    # via beautifulsoup4
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
